+++ CHANGELOG.md
@@ -30,6 +30,16 @@
 * ForkJoinPool threads owned by a ServiceHost are named like the host uri +
   the default name.
 
+* Add auto backup
+  When auto-backup is enabled, host performs incremental backup when underlying
+  document-index service persists document changes.
+  (for default `LuceneDocumentIndexService` service, it performs commit periodically)
+  The location is configurable by `autoBackup` host argument parameter.
+  To enable auto-backup, `isAutoBackupEnabled` argument is added for host start up.
+  Also, `/core/management` takes `AutoBackupConfiguration` patch request to toggle the
+  feature at run time.
+
+
 ## 1.5.2
 
 * "offset" parameter is added to QueryTask. This enables pagination logic to